# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dedicated API for ArcheryHub Mobile App (Archer \u0026 Scorekeeper)",
        "title": "ArcheryHub Mobile API",
        "termsOfService": "http://archeryhub.id/terms/",
        "contact": {
            "name": "ArcheryHub Support",
            "url": "https://archeryhub.id",
            "email": "support@archeryhub.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "host": "api.archeryhub.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/mobile/archer/events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get list of all events the archer is registered in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Archer"
                ],
                "summary": "Get My Events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileMyEventsResponse"
                        }
                    }
                }
            }
        },
        "/mobile/archer/events/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register the authenticated archer for one or more event categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Archer"
                ],
                "summary": "Register for Event",
                "parameters": [
                    {
                        "description": "Registration Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileEventRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileRegisterEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/mobile/archer/events/{id}/qr": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the digital ID / QR code for event check-in",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Archer"
                ],
                "summary": "Get Event QR Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/archer/events/{id}/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register the authenticated archer for one or more event categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Archer"
                ],
                "summary": "Register for Event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Registration Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileEventRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileRegisterEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/archer/events/{id}/registration": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get current user's registration and payment status for an event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Archer"
                ],
                "summary": "Get My Registration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Slug or UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileMyRegistrationResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mobile.MobileEventRegistrationRequest": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string",
                    "example": "arc-a49ee7d7-9d7b-4be7-8652-342f2fca23f9"
                },
                "event_category_id": {
                    "type": "string",
                    "example": "cat-recurve-adult-putra"
                },
                "event_category_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"cat-recurve-adult-putra\"",
                        " \"cat-recurve-adult-team\"]"
                    ]
                },
                "event_id": {
                    "type": "string",
                    "example": "national-open-2026"
                },
                "payment_amount": {
                    "type": "number",
                    "example": 450000
                },
                "payment_proof_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"https://cdn.archeryhub.id/media/proofs/receipt-001.jpg\"]"
                    ]
                },
                "payment_status": {
                    "type": "string",
                    "example": "menunggu acc"
                },
                "payment_type": {
                    "description": "manual, online, gateway",
                    "type": "string",
                    "example": "manual"
                },
                "registration_source": {
                    "type": "string",
                    "example": "mobile_app"
                }
            }
        },
...
        "mobile.MobileMyEventItem": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_slug": {
                    "type": "string"
                },
                "event_uuid": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_code_data_url": {
                    "type": "string"
                },
                "qr_raw": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "mobile.MobileMyEventsResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mobile.MobileMyEventItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "mobile.MobileMyRegistrationResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "registrations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mobile.MobileRegistrationItem"
                    }
                }
            }
        },
...
        "mobile.MobileRegisterEventResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "registered_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "registration_id": {
                    "type": "string"
                }
            }
        },
        "mobile.MobileRegistrationItem": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "checkout_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "pay_code": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_code_data_url": {
                    "type": "string"
                },
                "qr_raw": {
                    "type": "string"
                },
                "qr_url": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "tripay_reference": {
                    "type": "string"
                },
                "va_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}